move access location arrays
